Rename and move StickyLiveblogAsk component into marketing
